Almost final commit

See merge request informatik/qe/swa_swe/swa/swaws25/group1/g1t2!38

+++ swa-skeleton-main/src/main/frontend/src/components/RatingComponent.tsx
@@ -1,24 +1,12 @@
 import React, {useEffect, useMemo, useRef, useState} from "react";
-import {RatingDTO, RatingScale, ratingScaleToNumber, RatingTypes} from "../DTO/rating.types";
+import {numberToRatingScale, RatingDTO, ratingScaleToNumber, RatingTypes} from "../DTO/rating.types";
 import {RatingApi} from "../utilities/ratingApi";
 import {createRatingFromInterfaces} from "../utilities/ratingUtilities";
 import {UserxApi} from "../utilities/userxApi";
 import {UserxTypes} from "../DTO/userx.types";
-import {Divider} from "primereact/divider";
-import {Rating} from "primereact/rating";
-import {Tag} from "primereact/tag";
-import {FilterService} from "primereact/api";
 import RatingForm from "./RatingForm";
 import {Toast} from "primereact/toast";
 import RatingListComponent from "./RatingListComponent";
-
-FilterService.register('custom_range', (value, filters) => {
-    const [from, to] = filters ?? [null, null];
-    if (from === null && to === null) return true;
-    if (from !== null && to === null) return from <= value;
-    if (from === null && to !== null) return value <= to;
-    return from <= value && value <= to;
-});
 
 interface RatingComponentProps {
     productId: number;
@@ -29,11 +17,33 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
     const [ratingValue, setRatingValue] = useState<number | undefined>();
     const [comment, setComment] = useState<string>('');
     const [loading, setLoading] = useState<boolean>(true);
-    const [selectedRating, setRating] = useState<RatingDTO>(RatingTypes.empty);
+    const [selectedRating, setRating] = useState<RatingDTO>(RatingTypes.empty());
     const [sorting, setSorting] = useState<'asc' | 'desc'>('desc')
     const [ratingFilter, setRatingFilter] = useState<number | undefined>(undefined);
     const toast = useRef<Toast>(null)
+    const [user, setUser] = useState<UserxTypes | null>(null)
 
+    /**
+     * Determine if the user is a guest (This is a bit hacky but works)
+     */
+    useEffect(() => {
+        const determineUser = async () => {
+            try {
+                const authenticated : boolean = await UserxApi.isAuthenticated();
+                if (authenticated) {
+                    const author: UserxTypes = await UserxApi.getCurrentUser();
+                    setUser(author);
+                }
+            } catch (err: any) {
+                console.log("Error determining user: ", err);
+            }
+        }
+        void determineUser();
+    }, []);
+
+    /**
+     * Fetch all ratings from the backend to display them in the frontend
+     */
     useEffect(() => {
         const fetchAllRatingsForProduct = async () => {
             try {
@@ -52,35 +62,49 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
     const sortedRatings = useMemo(() => {
         return [...ratings]
             .filter(rating => {
-                // Filter nur anwenden, wenn ratingFilter gesetzt ist
                 if (ratingFilter !== undefined) {
-                    return ratingScaleToNumber(rating.rating) === ratingFilter;
+                    return ratingScaleToNumber(rating.ratingScale) === ratingFilter;
                 }
-                return true; // sonst alle Ratings behalten
+                return true;
             })
             .sort((a, b) =>
                 sorting === "asc"
-                    ? ratingScaleToNumber(a.rating) - ratingScaleToNumber(b.rating)
-                    : ratingScaleToNumber(b.rating) - ratingScaleToNumber(a.rating)
+                    ? ratingScaleToNumber(a.ratingScale) - ratingScaleToNumber(b.ratingScale)
+                    : ratingScaleToNumber(b.ratingScale) - ratingScaleToNumber(a.ratingScale)
             );
     }, [ratings, sorting, ratingFilter]);
 
+    /**
+     * Load the already written rating from the user if it exists.
+     * The rating form also acts as a way to update the rating later
+     */
     useEffect(() => {
-        if (loading) return;
         const loadUserRating = async () => {
-            const author: UserxTypes = await UserxApi.getCurrentUser();
-            const existingRating = ratings.find(rating => rating.authorId === author.id);
-
-            if (existingRating) {
-                setRating(existingRating);
-            } else {
-                setRating(RatingTypes.empty());
+            if (user != null) {
+                const existingRating = ratings.find(rating => rating.authorId === user.id);
+                if (existingRating) {
+                    setRating(existingRating);
+                } else {
+                    setRating(RatingTypes.empty());
+                }
             }
         }
         void loadUserRating();
-    }, [ratings, productId, loading]);
+    }, [user, ratings, productId]);
 
+    /**
+     * Create rating logic with edge case handling
+     */
     const createRating = async () => {
+        if (user == null) {
+            toast.current?.show({
+                severity: 'warn',
+                summary: 'You need an account to create a rating!',
+                detail: 'Click on Login to sign up now!',
+                life: 3000
+            });
+            return;
+        }
         if (ratingValue == undefined || comment == '') {
             toast.current?.show({
                 severity: 'warn',
@@ -93,7 +117,7 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
         const author: UserxTypes = await UserxApi.getCurrentUser();
         const ratingToSave = new RatingTypes({
             ...selectedRating,
-            rating: numberToRatingScale(ratingValue),
+            ratingScale: numberToRatingScale(ratingValue),
             comment: comment,
             authorId: author.id,
             username: author.username,
@@ -105,6 +129,9 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
         setRatings(prevState => [...prevState, newRating])
     };
 
+    /**
+     * Update rating logic with edge cases. This option only gets executed if the update Button is pressed.
+     */
     const updateRating = async () => {
         if (ratingValue == undefined || comment == '') {
             toast.current?.show({
@@ -121,7 +148,7 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
         try {
             const ratingToUpdate: RatingDTO = {
                 ...selectedRating,
-                rating: numberToRatingScale(ratingValue),
+                ratingScale: numberToRatingScale(ratingValue),
                 comment: comment,
                 authorId: selectedRating.authorId,
                 username: selectedRating.username,
@@ -136,12 +163,11 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
         }
     }
 
-    const onEditRating = (rating: RatingTypes) => {
-        setRating(rating);
-    }
-
+    /**
+     * Delete logic with edge case handling. Only gets executed with the delete button
+     */
     const deleteRating = async () => {
-        if (!selectedRating.id) return;
+        if (!user || !selectedRating.id) return;
 
         try {
             await RatingApi.deleteRating(productId, selectedRating);
@@ -152,23 +178,9 @@ const RatingComponent: React.FC<RatingComponentProps> = ({productId}) => {
         }
     }
 
-    const numberToRatingScale = (ratingValue: number | undefined): RatingScale | undefined => {
-        switch (ratingValue) {
-            case 1:
-                return RatingScale.ONE_STAR;
-            case 2:
-                return RatingScale.TWO_STARS;
-            case 3:
-                return RatingScale.THREE_STARS;
-            case 4:
-                return RatingScale.FOUR_STARS;
-            case 5:
-                return RatingScale.FIVE_STARS;
-            default:
-                return undefined;
-        }
-    }
-
+    /**
+     * Rating component using RatingForm and RatingListComponent
+     */
     return (
         <div>
             <Toast ref={toast} position="top-right"/>
